pyproject.toml
# pyproject.toml

[build-system]
requires      = [
    "setuptools>=65.5.0",
    "flask>=2.2.3",
    "flask-sqlalchemy>=3.0.3",
    "flask-migrate>=4.0.4",
    "flask-httpauth>=4.8.0",
    "itsdangerous>=2.1.2",
    "psycopg2-binary>=2.9.5",
    "sqlalchemy>=2.0.7",
]
build-backend = "setuptools.build_meta"

[project]
name = "rss"
version = "1.0.0"
description = "API for the rss website"
readme = "README.md"
authors = [{ name = "Daniel Ibarrola", email = "daniel.ibarrola.sanchez@gmail.com" }]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["earthquake", "alerts", "cap", "rss"]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["bumpver", "pytest", "pytest-mock", "flask-cors"]

[project.urls]
Homepage = "https://github.com/Daniel-Ibarrola/RSS.git"


[tool.bumpver]
current_version = "1.0.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]
